expand v2 is done...for now
