+++ .gitignore
@@ -15,3 +15,4 @@ a.out
 info.plist
 test1
 minishell
+expand.c
