+++ Makefile
@@ -26,8 +26,7 @@ CC			=	cc
 CFILES		=	src/input_parsing.c		src/freeing_utils.c		src/built_ins.c		src/echo.c		\
 				src/pipex.c				src/fd_utils.c			src/check_redirs.c	src/forking.c	\
 				src/cmd_parsing.c		src/list_utils.c		src/get_redirs.c	src/main.c		\
-				src/utils.c				src/export.c									\
-				src/ft_memcpy_array.c	src/expand_v2.c
+				src/utils.c				src/export.c			src/expand_v2.c
 
 OFILES		=	$(CFILES:.c=.o)
